Game repository, service, controller

+++ src/main/java/com/omarcosallan/dslist/entities/Game.java
@@ -10,7 +10,7 @@ public class Game {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String title;
 
     @Column(name = "game_year")
@@ -28,7 +28,7 @@ public class Game {
 
     public Game() {}
 
-    public Game(Integer id, String title, Integer year, String genre, String platforms, Double score, String imgUrl, String shortDescription, String longDescription) {
+    public Game(Long id, String title, Integer year, String genre, String platforms, Double score, String imgUrl, String shortDescription, String longDescription) {
         this.id = id;
         this.title = title;
         this.year = year;
@@ -40,11 +40,11 @@ public class Game {
         this.longDescription = longDescription;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
